feat: Added Templates.

+++ index.js
@@ -3,10 +3,8 @@
 const { execSync } = require('child_process');
 const fs = require('fs-extra');
 const path = require('path');
-const inquirer = require('inquirer');
 
 (async () => {
-    // Dynamically import chalk
     const chalk = (await import('chalk')).default;
 
     // Check if yarn is installed
@@ -20,35 +18,38 @@ const inquirer = require('inquirer');
     }
 
     const projectName = process.argv[2];
+    const templateName = "template-1";
+
     if (!projectName) {
         console.error(chalk.red('Please provide a project name.'));
         process.exit(1);
     }
 
     const targetPath = path.join(process.cwd(), projectName);
+    const templatePath = path.join(__dirname, 'templates', templateName);
 
-    // Prompt the user to select a template
-
-    const templatePath = path.join(__dirname, 'templates', 'template-1');
+    if (!fs.existsSync(templatePath)) {
+        console.error(chalk.red(`Template "${templateName}" does not exist.`));
+        process.exit(1);
+    }
 
     // Copy the selected template to the target directory
     try {
         await fs.copy(templatePath, targetPath);
-        console.log(chalk.green(`Template ${template} copied to ${targetPath}`));
+        console.log(chalk.green(`Template "${templateName}" copied to ${targetPath}`));
     } catch (error) {
         console.error(chalk.red('Error copying template:', error));
         process.exit(1);
     }
 
-    // Change directory to the target path
     process.chdir(targetPath);
 
     // Install required packages
     console.log(chalk.blue('Installing packages...'));
-    execSync('yarn install', { stdio: 'inherit' }); // or 'npm install'
+    execSync('yarn install', { stdio: 'inherit' });
 
     console.log(chalk.green(`Project ${projectName} is ready!`));
     console.log(chalk.yellow(`To start working on your project, run:`));
     console.log(chalk.cyan(`\t cd ${projectName}`));
-    console.log(chalk.cyan('\t yarn dev')); // or any other command you want to suggest
+    console.log(chalk.cyan('\t yarn dev'));
 })();
